requirements.txt
requests>=2.25.1
prettytable>=2.0.0
colorama>=0.4.4
pyyaml>=5.4.1 